requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor>=3.3.0
jinja2==3.1.2
python-dotenv==1.0.0
openai>=1.0.0
pandas==2.1.4
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
# Additional dependencies for better functionality
httpx>=0.25.0
pymongo>=4.5.0,<5.0.0
bcrypt==4.0.1
cryptography>=41.0.0
# Optional: for better development experience
python-multipart>=0.0.6
aiofiles>=23.0.0 